add some support for classes rather than in-line styles
i.e. track/recognise that a color was one of the 16 or 256, and output
the proper class name for it (fore and background); and use class names
for "features" (i.e. bold, bright, etc) if enabled.

Also, provide an option to output the contents of the style tag for the
chosen palette.

Also, "fix" the setting of the background/foreground to "initial",
unneeded.