repo: opt-in/default x86_64 AVX2/SSE2 if available
- Show the "status" of whether "-S" is running AVX2, SSE2 or the "old"
  implementation (if opted out via WANT_DEFAULT).
- Opt into via WANT_AVX2, WANT_SSE2, WANT_DEFAULT or let the compiler
  pick.
- Workflows peppered with "--help" to show what the flow got. The build
  and tags don't have -march=native, so they shouldn't build AVX2 under
  x86_64. Note you'll get SSE2 on x86_64 as every x86_64 can do SSE2.
  But if you locally build it via "make", you'll likely have AVX2 if you
  add "-march=native" and you've a AVX2-capable processor.

What the "improved" versions of "-S" and ansi2html do is go through the
input buffer, "as usual", but first try to see if we're "lucky" and are
in the state where we're handling "text" and looking for an ESC
character (only, for stripping; or also "&", "<" and ">" for ansi2html),
and just so happen to have a 16- (for SSE2) or 32-bytes (for AVX2)
series of bytes which are NOT in that list. If so, we can just (output
the span if we need to, for ansi2html) copy the lot to the output
buffer, without any need to go through the state machine byte by byte,
and maintaining the "STATE_TEXT" state, too.

To speed things up a little bit further, this is only performed if we're
at the "right" *boundary* to use _aligned_ AVX2/SSE2 instructions.

If an ESC (or funny character) is found, and it's at a byte greater than
0, we "shortcut" by memcpy'ing the bytes between the first and the
non-text character to the output buffer (no need to get through the
state machine for that!), and then proceed with the "usual" processing
byte by byte from the ESC (or funny character) on.

For a "silly" case where the whole input is large and doesn't have any
escape character at all (i.e. say it's a 1GiB of "just text", no ESC at
all), this speeds things up fairly considerably (on my machine, for a
1.1GiB of "just text", from 1.5s to around 150ms for AVX2 and 190ms for
SSE2).

In the most pernicious case where the whole input is pretty much mostly
escape codes, i.e. repetitions of "\e[0;38;5;NNNmNNN" or similar, and
therefore the chance to hit a streak of "all non-esc" is pretty much
nil, the performance is a wee bit worse than previously, but that's not
really a common enough case IMO.

For my usual 500MiB test file with a mix and match of heavy ANSI code
and a bit of text, this version can be a wee bit faster: around 390ms
for the AVX2 version, vs 430+ms for the older version.